+++ test/test_preproc.py
@@ -29,12 +29,21 @@ def test_hover():
     root_dir = os.path.join(test_dir, "pp")
     string = write_rpc_request(1, "initialize", {"rootPath": root_dir})
     file_path = os.path.join(test_dir, "pp", "preproc.F90")
-    string += hover_req(file_path, 5, 8)
-    string += hover_req(file_path, 7, 30)
+    string += hover_req(file_path, 5, 8)  # user defined type
+    string += hover_req(file_path, 7, 30)  # variable
+    string += hover_req(file_path, 7, 40)  # multi-lin variable
+    string += hover_req(file_path, 8, 7)  # function with if conditional
+    string += hover_req(file_path, 9, 7)  # multiline function with if conditional
     errcode, results = run_request(string, f" --config={root_dir}/.pp_conf.json")
     assert errcode == 0
 
     # Reference solution
-    ref_results = ("#define PCType character*(80)", "#define PETSC_ERR_INT_OVERFLOW 84")
+    ref_results = (
+        "#define PCType character*(80)",
+        "#define PETSC_ERR_INT_OVERFLOW 84",
+        "#define varVar 55",
+        "#define ewrite if (priority <= 3) write((priority), format)",
+        "#define ewrite2 if (priority <= 3) write((priority), format)",
+    )
     assert len(ref_results) == len(results) - 1
     check_return(results[1:], ref_results)